+++ R/mcmc-traces.R
@@ -66,13 +66,11 @@
 #'   }
 #'   \item{`mcmc_rank_ecdf()`}{
 #'    The ECDFs of the ranks from `mcmc_rank_hist()` are plotted with the
-#'    simultaneous confidence bands with a coverage determined by `prob`,
-#'    that is, bands that completely cover all of the rank ECDFs with the
-#'    probability 'prob'.
-#'    By default, the difference between the observed rank ECDFs and the
-#'    theoretical expectation for samples originating from the same distribution
-#'    is drawn.
-#'    See Säilynoja et al. (2021) for details.
+#'    simultaneous confidence bands with a coverage determined by `prob`, that
+#'    is, bands that completely cover all of the rank ECDFs with the probability
+#'    `prob`. If `plot_diff = TRUE`, the difference between the observed rank
+#'    ECDFs and the theoretical expectation for samples originating from the
+#'    same distribution is drawn. See Säilynoja et al. (2021) for details.
 #'   }
 #' }
 #'
@@ -117,10 +115,11 @@
 #' mcmc_rank_hist(x, pars = c("alpha", "sigma"), ref_line = TRUE)
 #' mcmc_rank_overlay(x, "alpha")
 #'
-#' # ECDF difference plots of the ranking of MCMC samples between chains.
+#' # ECDF and ECDF difference plots of the ranking of MCMC samples between chains.
 #' # Provide 99% simultaneous confidence intervals for the chains sampling from
 #' # the same distribution.
 #' mcmc_rank_ecdf(x, prob = 0.99)
+#' mcmc_rank_ecdf(x, prob = 0.99, plot_diff = TRUE)
 #'
 #' \dontrun{
 #' # parse facet label text
@@ -306,7 +305,7 @@ mcmc_rank_overlay <- function(x,
     group_by(.data$cut) %>%
     mutate(bin_start = min(.data$value_rank)) %>%
     ungroup() %>%
-    select(-.data$cut)
+    select(-c("cut"))
 
   d_bin_counts <- data %>%
     left_join(histobins, by = "value_rank") %>%
@@ -342,7 +341,7 @@ mcmc_rank_overlay <- function(x,
   }
 
   ggplot(d_bin_counts) +
-    aes_(x = ~ bin_start, y =  ~ n, color = ~ chain) +
+    aes(x = .data$bin_start, y =  .data$n, color = .data$chain) +
     geom_step() +
     layer_ref_line +
     facet_call +
@@ -415,13 +414,13 @@ mcmc_rank_hist <- function(x,
   facet_call <- do.call(facet_f, facet_args)
 
   ggplot(data) +
-    aes_(x = ~ value_rank) +
+    aes(x = .data$value_rank) +
     geom_histogram(
       color = get_color("mid_highlight"),
       fill = get_color("mid"),
       binwidth = right_edge / n_bins,
       boundary = right_edge,
-      size = .25
+      linewidth = 0.25
     ) +
     layer_ref_line +
     geom_blank(data = data_boundaries) +
@@ -442,7 +441,7 @@ mcmc_rank_hist <- function(x,
 #' @param prob For `mcmc_rank_ecdf()`, a value between 0 and 1
 #' specifying the desired simultaneous confidence of the confidence bands to be
 #' drawn for the rank ECDF plots.
-#' @param plot_diff For `mcmc_rank_ecdf()`, a boolean specifying it the
+#' @param plot_diff For `mcmc_rank_ecdf()`, a boolean specifying if the
 #' difference between the observed rank ECDFs and the theoretical expectation
 #' should be drawn instead of the unmodified rank ECDF plots.
 #' @export
@@ -455,8 +454,8 @@ mcmc_rank_ecdf <-
            K = NULL,
            facet_args = list(),
            prob = 0.99,
-           plot_diff = TRUE,
-           interpolate_adj = TRUE) {
+           plot_diff = FALSE,
+           interpolate_adj = NULL) {
   check_ignored_arguments(...,
     ok_args = c("K", "pit", "prob", "plot_diff", "interpolate_adj", "M")
   )
@@ -504,7 +503,7 @@ mcmc_rank_ecdf <-
     x = x
   )
   data <- data %>%
-    group_by(parameter, chain) %>%
+    group_by(.data$parameter, .data$chain) %>%
     dplyr::group_map(~ data.frame(
       parameter = .y[1],
       chain = .y[2],
@@ -514,11 +513,11 @@ mcmc_rank_ecdf <-
     )) %>%
     dplyr::bind_rows()
 
-  mapping <- aes_(
-    x = ~x,
-    y = ~ecdf_value,
-    color = ~chain,
-    group = ~chain
+  mapping <- aes(
+    x = .data$x,
+    y = .data$ecdf_value,
+    color = .data$chain,
+    group = .data$chain
   )
 
   scale_color <- scale_color_manual("Chain", values = chain_colors(n_chain))
@@ -533,8 +532,8 @@ mcmc_rank_ecdf <-
   }
 
   ggplot() +
-    geom_step(data = data_lim, aes_(x = ~x, y = ~upper), show.legend = FALSE) +
-    geom_step(data = data_lim, aes_(x = ~x, y = ~lower), show.legend = FALSE) +
+    geom_step(data = data_lim, aes(x = .data$x, y = .data$upper), show.legend = FALSE) +
+    geom_step(data = data_lim, aes(x = .data$x, y = .data$lower), show.legend = FALSE) +
     geom_step(mapping, data) +
     bayesplot_theme_get() +
     scale_color +
@@ -640,23 +639,23 @@ mcmc_trace_data <- function(x,
   n_chain <- unique(data$n_chains)
   n_param <- unique(data$n_parameters)
 
-  mapping <- aes_(
-    x = ~ iteration,
-    y = ~ value,
-    color = ~ chain
+  mapping <- aes(
+    x = .data$iteration,
+    y = .data$value,
+    color = .data$chain
   )
 
   if (!is.null(highlight)) {
-    mapping <- modify_aes_(
+    mapping <- modify_aes(
       mapping,
-      alpha = ~ highlight,
-      color = ~ highlight
+      alpha = .data$highlight,
+      color = .data$highlight
     )
   }
 
   layer_warmup <- if (n_warmup > 0) {
     layer_warmup <- annotate(
-      "rect", xmin = -Inf, xmax = n_warmup, ymin = -Inf, ymax = Inf, size = 1,
+      "rect", xmin = -Inf, xmax = n_warmup, ymin = -Inf, ymax = Inf, linewidth = 1,
       color = "gray88", fill = "gray88", alpha = 0.5
     )
   } else {
@@ -793,13 +792,13 @@ divergence_rug <- function(np, np_style, n_iter, n_chain) {
   }
 
   geom_rug(
-    aes_(x = ~ Divergent, linetype = "Divergence"),
+    aes(x = .data$Divergent, linetype = "Divergence"),
     data = div_info,
     na.rm = TRUE,
     inherit.aes = FALSE,
     sides = "b",
     color = np_style$color[["div"]],
-    size = np_style$size[["div"]],
+    linewidth = np_style$size[["div"]],
     alpha = np_style$alpha[["div"]]
   )
 }