+++ NEWS.md
@@ -1,8 +1,19 @@
 <!-- See http://style.tidyverse.org/news.html for advice on writing news -->
 
-# bayesplot 1.9.0.9000
+# bayesplot 1.10.0
 
-Items for next release go here
+* New function `mcmc_rank_ecdf()` for rank ecdf plots with confidence bands for
+assessing if two or more chains sample the same distribution (#282,
+@TeemuSailynoja)
+* New functions `ppc_pit_ecdf()`, `ppc_pit_ecdf_grouped()`, PIT ecdf plots with
+confidence bands to assess if `y` and `yrep` contain samples from the same
+distribution. (#282, @TeemuSailynoja)
+* Several `ppc` and `ppd` functions now accept the new `linewidth` argument
+introduced in ggplot2 3.4.0: `ppc_bars()`, `ppc_bars_grouped()`,
+`ppc_intervals()`, `ppc_intervals_grouped()`, `ppd_intervals()`,
+`ppd_intervals_grouped()`.
+* Fix bug in how `mcmc_pairs()` detected hitting `max_treedepth`, thanks to @dmphillippo. (#281)
+* Fix failing tests due to changes in ggplot2 3.4.0 (#289)
 
 # bayesplot 1.9.0